Create README.md
Project developed by MARTINI DAVIDE and NIGRO SIMONE for "Bioinformatics Algorithms" course. The professor in charge of this course is Matteo Comin of University of Padua.
The goal of this project is to analyze and compare run time, and memory usage from KMC, Turtle and DSK, for the count of k-mers in FASTA dataset.
FASTA dataset is created by a function in 'analyze.c' file. See 'FASTA.gen.c' for more information.
We developed a brute force algorithm to find and count, k-mers in dataset. You can see this algorithm in 'BFKC.c' file.
In 'Data' folder, you can find '*.csv' that we created for compare this tools.
For more information and the description of the steps of this project read the paper "Algoritmi per la Bioinformatica.pdf"